refactor(streams): Move BaseDuplexStream to own file
