+++ packages/streams/src/ChromeRuntimeStream.ts
@@ -17,8 +17,9 @@
 import type { Json } from '@metamask/utils';
 import { stringify } from '@ocap/utils';
 
+import { BaseDuplexStream } from './BaseDuplexStream.js';
 import type { OnEnd, ReceiveInput } from './BaseStream.js';
-import { BaseDuplexStream, BaseReader, BaseWriter } from './BaseStream.js';
+import { BaseReader, BaseWriter } from './BaseStream.js';
 import type { ChromeRuntime, ChromeMessageSender } from './chrome.js';
 import type { Dispatchable } from './utils.js';
