+++ packages/streams/src/MessagePortStream.ts
@@ -21,8 +21,9 @@
 
 import type { Json } from '@metamask/utils';
 
+import { BaseDuplexStream } from './BaseDuplexStream.js';
 import type { OnEnd } from './BaseStream.js';
-import { BaseDuplexStream, BaseReader, BaseWriter } from './BaseStream.js';
+import { BaseReader, BaseWriter } from './BaseStream.js';
 import type { Dispatchable } from './utils.js';
 
 /**
