+++ packages/streams/tsconfig.json
@@ -6,5 +6,5 @@
     "types": ["ses", "vitest", "vitest/jsdom"]
   },
   "references": [{ "path": "../test-utils" }],
-  "include": ["./src", "./test/envelope-kit-fixtures.ts", "./src/chrome.d.ts"]
+  "include": ["./src", "./src/chrome.d.ts", "./test/*.ts"]
 }
